l2tp: marshal quiescent tunnel close via. transport reader goroutine
When closing the quiescent tunnel it's important to first ensure
that the tunnel's goroutine has terminated prior to closing the
transport.

Failure to do so results in the transport close triggering a read
error.  If the quiescent tunnel tears itself down on a read error
we end up trying to close the transport twice, leading to a panic.