l2tp/l2tp: allow local address to be unset
When creating a dynamic tunnel we don't necessarily want to specify
the local address, for two reasons.

Firstly, leaving the local address unspecified allows the kernel to
figure out the correct address to use based on the routing table.

Secondly, leaving the local address unspecified allows the kernel to
allocate a port, minimising the chance of UDP port clashes.