l2tp/avp: fix up encoding of variable-length data
binary.Write only handles fixed-size types or slices of fixed-size
types.

Since a string is neither of these, it's not possible to directly
write a string using binary.Write.  This means we cannot directly
write string or resultCode (which contains a string) using
binary.Write.

Refactor to account for this.