l2tp/quiescent: avoid shutdown race
Because the tunnel goroutine can exit of its own accord, trying to
close via. the context by sending a message on a channel can result
in hangs if the goroutine has gone by the point that Close() is called.
This is because nothing is reading the channel, and hence the sender
blocks.

To avoid this problem, signal shutdown by closing the channel.  This
works just as well to wake up the tunnel goroutine, with the added
advantage that close() doesn't block even if the goroutine has gone
away.

This does of course mean that Close() can be called once only!