l2tp: add tunnel config option for stopccn pend timeout
RFC2661 section 5.7 says that on receipt of a StopCCN message, the
receiver should persist for a full retransmit period in case the ACK
for the StopCCN is lost and the peer retransmits.

By default it recommends a full retransmit period of 31s.

Add a configuration knob to control this, and default to 31s.