l2tp/transport: refactor to separate receive and transmit paths
Previously the transport had a single go routine running the bulk of
the transport logic, and being fed received messages from a seperate
go routine reading from the control plane socket.

This design was inherently racy: if a tunnel happens to send a message
while the transport is in the process of receiving a message you end
up with the tunnel go routine blocking in the channel send to the
transport, and the transport go routine blocked in the channel send
to the tunnel.

To avoid this, split transmit and receive into separate go routines.

The majority of the transport logic is handled by the transmit go
routine.  The receive go routine processes received frames and sends
them in sequence on the receive channel.  Sequence number updates are
forwarded to the transmit go routine for processing the ack queue and
toggling timers.