l2tp/tunnel: synchronise session addition
If a tunnel has a session added while it's sending SCCRQ, the
injectEvent call will block waiting for control to return to the
tunnel's main event loop.

If the tunnel never returns there (if SCCRQ times out and the transport
dies), then the session addition will block for ever and an application
can hang.

To avoid this, synchronise NewSession around the tunnel isClosing flag,
protected by a mutex.

TODO: there may be a better way to address this.  An application would
block in NewSession waiting for the channel write to complete, which may
not be a good idea.