kpppoed: use nil netlink interface for unit tests where we're just checking PPPoE packet exchanges
In order to make this work we need to slightly tweak the test client
to avoid a deadlock.

Because the nil l2tpd implementation wait function returns immediately
l2tpd appears to exit immediately.  This causes kpppoed to send a PADT
packet after the PADS; which then causes the test client to block trying
to forward the PADT on the receive channel.

In order to avoid this, use a non-blocking channel send for the test
client.