kl2tpd: add an interface for abstracting pseudowire instances
Different session types require the L2TP daemon to do different things
to complete bringup of the pseudowire.

For example, a PPP session requires pppd launching to terminate the PPP
session carried by the L2TP session.

A PPPAC session requires pppox sockets to be created in order to bridge
the pppoe and pppol2tp channels in the kernel ppp subsystem.

Add an interface to abstract these details somewhat in the kl2tpd code.