config: fix up TestGetTunnels
TestGetTunnels was sometimes failing because the ordering of named
session instances inside each tunnel isn't deterministic, hence:

  if !reflect.DeepEqual(got, &want) {
    t.Fatalf("got %v, want %v", got, want)
  }

would erratically fail if the expected ordering didn't match the parsed
output.

There's not any reason for the named session instances to follow any
particular order in the parsed data, so it makes most sense to fix this
by relaxing the test code's expectations.

This change modifies the test to validate that each tunnel contains the
expected sessions, and that each session contains the expected
configuration parameters.