l2tp: fix up transport test when kernel doesn't support IP encap
The transport tests can generally be run without root permissions,
assuming that the kernel has the l2tp_ip and l2tp_ip6 modules loaded
for IP encapsulation support.

On a new Fedora 38 system, the l2tp_ip module is blacklisted to prevent
automatic loading by non-root users[1], which causes the first IP
encapsulation test to fail.

Slightly ironically, the l2tp_ip6 module is /not/ blacklisted, so the
subsequent IP6 encapsulation test will cause both l2tp_ip6 and l2tp_ip
to load (the former depends on the latter), and the IP encap tests will
pass thereafter.

For developers this is probably not so much of an issue (we can expect
them to load l2tp modules), but for automated package builds the
underlying host may not have IP encap support, potentially leading to
package build failures.

As such, work around the test failure by skipping the unit test when
the kernel lacks protocol support.

[1].  This seems to be driven by RHEL policy -- there is a document
which details the background of the decision although it is paywalled:

https://access.redhat.com/articles/3760101