l2tp/avp: don't reject L2TPv2 AVPs
There are some lesser-used L2TPv2 AVPs which go-l2tp doesn't currently
do anything meaningful with.

However, if we do receive one of these AVPs, we do want to be able to at
least parse it: if not, we'll potentially end up tearing the tunnel down
if the AVP is flagged as mandatory by the peer.

Treat the following L2TPv2 AVPs as byte arrays:

 * Q.931 Cause Code (RFC2661 section 4.4.4)
 * Call Errors (RFC2661 section 4.4.6)
 * ACCM (RFC2661 section 4.4.6)

All of these AVPs have some degree of internal structure which the
parser can (and at some point should) learn to understand.

For the time being, the avpDataTypeBytes type implies a minimum of
processing by the parsing code, but allows the parser to handle the
AVPs without falling over.

While we're here, treat the draft-only Packet Proc Delay in the same
way: it's unlikely we'll see it in the wild, but you never know.

Finally, update the comment for the L2TPv3-only Pseudowire Caps AVP,
which can stay as avpDataTypeUnimplemented since go-l2tp doesn't
currently implement the v3 control plane.