l2tp/tunnel: add FSM handling for SLI and WEN messages
We don't currently do anything useful with these messages, just log
them (it's worth noting xl2tpd takes the same approach with SLI).

Interestingly, while RFC2661 defines these control messages it doesn't
call them out explicitly in the FSM tables, so it's not clear when they
should be expected.

However from the context it makes sense they'd be handled after the
tunnel and session was established.

For the time being, since we're just ignoring the messages, there's
little point plumbing them through to the session, so this commit
doesn't do that.

Set-Link-Info might be sent by the LNS to the LAC to set PPP-negotiated
options.

RFC2661 states:

        These options can change at any time during the life
        of the call, thus the LAC MUST be able to update its internal
        call information and behavior on an active PPP session.

In order to do this, go-l2tp would need some way to dynamically send
updated ACCM information to pppd, which implies some kind of IPC between
pppd and go-l2tp.

WAN-Error-Notify shouldn't really be seen by go-l2tp currently since
it's supposed to be sent by the LAC to the LNS.  However we might as
well add minimal handling for it just in case.

Fixes #6