l2tp/msg: log and ignore data packets received over the control socket
When bringing up a session as a client, go-l2tp creates the dataplane
for that session once the controlplane has fully established.

The controlplane is considered to be fully established once go-l2tp's
ICCN message has been acknowledged by the server by the rules of the
L2TP reliable transport algorithm.

Depending on the implementation of the server, and the vagaries of UDP
transmission, it's possible that the control packet acknowledging the
ICCN might arrive after the first dataplane frame.

In this scenario, the Linux kernel doesn't know what to do with the
first data frame, because go-l2tp hasn't yet created the dataplane for
the session.  It defaults to passing the data frame up to userspace in
the hopes userspace will be able to do something with it.

As such, it's possible for go-l2tp's parsing code to see data frames.
It will typically reject these, although with a rather confusing error
message:

    level=error tunnel_name=t1 function=transport message="frame receive failed"
        error="malformed header: length 17024 exceeds buffer bounds of 30"

This error stems from the fact that the Linux kernel doesn't set the
length L2TP header field for data packets: that's perfectly legitimate,
but it results in a header with the tunnel ID at the offset the length
field would normally be for a control packet.

As a result, go-l2tp parses the tunnel ID as the packet length.  Since
tunnel IDs are often randomly generated it's likely that the tunnel ID
as a numeric value will be greater than the length of the frame on the
wire, and so go-l2tp will reject the frame early in the parsing process.

This generally isn't harmful, since the logging on a "malformed header"
is just informational: however it does confound debugging efforts.

However, in the edge case that the tunnel ID was less than or equal to
the length of the frame on the wire, go-l2tp wouldn't reject the frame
and would go ahead and try to (and fail to) parse it as a control frame.
This would almost certainly cause the tunnel to be torn down.

To avoid these problems, go-l2tp can check the message type bit in the
L2TP header prior to doing any further parsing.  That allows data frames
to be dropped earlier in the process with explicit logging.

ref: https://bbs.archlinux.org/viewtopic.php?pid=1970197#p1970197